refactor(dev): add more building entries
