+++ package.json
@@ -55,7 +55,10 @@
     "build": "webpack",
     "build-lib": "babel src --out-dir lib",
     "ci": "npm run lint && npm run test",
-    "compress": "uglifyjs -c -m -o dist/g2.min.js -- build/g2.js",
+    "compress-g2": "uglifyjs -c -m -o dist/g2.min.js -- build/g2.js",
+    "compress-core": "uglifyjs -c -m -o dist/g2-core.min.js -- build/g2-core.js",
+    "compress-simple": "uglifyjs -c -m -o dist/g2-simple.min.js -- build/g2-simple.js",
+    "compress": "npm run compress-g2 && npm run compress-core && npm run compress-simple",
     "coverage": "npm run coverage-generator && npm run coverage-viewer",
     "coverage-generator": "torch --compile --coverage --renderer --recursive test/unit",
     "coverage-viewer": "torch-coverage",
